+++ src/Text/LLVM/DebugUtils.hs
@@ -230,16 +230,28 @@ debugInfoToStructField mdMap di =
      let bitfield | testBit (didtFlags dt) 19
                   , Just extraData      <- didtExtraData dt
                   , Just bitfieldOffset <- getInteger mdMap extraData
-                  = Just $ BitfieldInfo { biFieldSize      = didtSize dt
-                                        , biBitfieldOffset = fromInteger bitfieldOffset
-                                        }
+                  = do size <- getSizeOrOffset (didtSize dt)
+                       Just $ BitfieldInfo { biFieldSize      = size
+                                           , biBitfieldOffset = fromInteger bitfieldOffset
+                                           }
                   | otherwise
                   = Nothing
+     offset <- getSizeOrOffset (didtOffset dt)
      Just (StructFieldInfo { sfiName     = fieldName
-                           , sfiOffset   = didtOffset dt
+                           , sfiOffset   = offset
                            , sfiBitfield = bitfield
                            , sfiInfo     = valMdToInfo' mdMap (didtBaseType dt)
                            })
+  where
+    -- TODO: Currently, this only recognizes bare integer (i.e., 'ValInteger')
+    -- sizes and offsets. This is likely good enough for Clang-derived LLVM, but
+    -- Ada-derived LLVM may contain more complex metadata values that this
+    -- currently doesn't handle.
+    getSizeOrOffset :: Maybe ValMd -> Maybe Word64
+    getSizeOrOffset (Just (ValMdValue tv))
+      | ValInteger i <- typedValue tv
+      = Just (fromInteger i)
+    getSizeOrOffset _ = Nothing
 
 
 getUnionFields :: MdMap -> DICompositeType -> Maybe [UnionFieldInfo]